requirements.txt
streamlit
pandas
PyGithub
plotly
python-dotenv
streamlit-option-menu